+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [Library] `password input` component (Orange-OpenSource/ouds-ios#997)
 - [Library] Add flag to let `button` takes the screen full width (Orange-OpenSource/ouds-ios#1252)
 - [Library] Stop animation of switch if defined in accessibility settings (Orange-OpenSource/ouds-ios#660)
 - [Library] Display in logs components versions